review: refactor
